Solid colored cube using backface culling.
Made a Vector prototype and fixed the normals on the faces array. Added a tumbling animation (initially to help debug backface culling) that should prove useful as I develop the more complex Rubik's Cube geometry. Fixed a transformation bug that caused the center to be off in rectangular canvases. Added some missing semi-colons.